游戏列表页增加gameType字段返回

+++ src/main/java/com/dx/facade/game/resp/GameOrderRecordUserVenueResp.java
@@ -31,7 +31,7 @@ public class GameOrderRecordUserVenueResp implements Serializable {
     * 游戏类型
     */
    @ApiModelProperty("游戏类型")
-   private Integer gameType;
+   private Long gameType;
    /**
     * 状态
     */
